renamed from ChartHighstock to Stock
